day03: simplify index array initialization
In retrospect, we don't lose _that_ much time 0-initializing a handful
of bytes in the array once per bank, and doing so makes the whole thing
smaller, simpler, and safer. The exact same logic which said that
it was safe to use an uninitialized array says that we know for a fact
we'll be overwriting every index of an initialized array, so
we may as well simplify.