day04: implement part 2
Would have worked first try, if only I'd remembered that just swapping
the two maps would undo the work and produce an infinite loop. Some
cloning is required, but not all that slow in the end.

The double-block while statement is kind of funky. I think I like it,
but what I really need is the me a decade from now to git blame, discover
this comment, and decide whether I do in fact like it.