day07: impl part 2
The good news is that this was honestly pretty straightforward:
took a little bit of thought to come up with the right approach,
but it was one of those nice situations where fundamentally we can
use the same main loop in both parts.

The bad news is that I can't take credit for any of it. While I
had previously been on Reddit trying to even understand what was
intended for part 1, I came across several memes, one of which
was "pascal's triangle: same logic". Which unfortunately trivialized
this whole thing.

(I have no idea what the person was doing who combined BFS and DFS
in this problem, who then made a meme about it. This is not in any
obvious way a search problem, to me.)