day08: impl part 1
Whoof, I feel kind of stupid. Got all wrapped around the axle trying
to even figure out how to count iterations, and then also managed to
fail to consider that joining two circuits would have to be able to
adjust more than a single item, if we ever join two circuits which
each have more than a single member.

Representing the circuit assignments as an array of IDs seems obvious
now that I'm looking at that answer, but I also had a weirdly hard
time coming up with that. Storing circuits as the set of points
contained within them was my first attempt, but that... was harder.

Got there in the end though. The ultimate measure is success.