day09: impl part 1
I was somehow both over and underthinking this: I had a whole system
set up for running a diagonal projection up to get the points which
necessarily would form the max area by only testing a few points...
and it didn't work. No idea why, don't quite understand it, I thought my
logic was valid, but yeah, nothing.

Then I realized that yesterday I took the cartesian product of a list
with itself, and the list was twice this size. So, I just had to do the
simple thing.