day10: impl part 1 (partial)
The classic: works for the example, fails for the input.

Input fails complaining that a particular machine can't be activated
no matter which buttons are pressed. No idea yet what's going wrong here.
Could be my button-pressing logic. Could be parsing. I can't yet
be bothered to figure it out. It's late.