day10: fix impl for day 1
Ultimately exactly one line of this change mattered, with regard to the
solution: I was accidentally generating permutations over a field as wide
as the number of lights, where I needed to be generating permutations
over a field as wide as the number of buttons. The failing input just
happened to be the first one which required a button to be pressed which
had a position wider than the number of lights.